+++ cupy/array_api/_constants.py
