+++ cupy/array_api/_data_type_functions.py
