+++ cupy/array_api/_searching_functions.py
