+++ cupy/array_api/_statistical_functions.py
