+++ setup.cfg
@@ -1,5 +1,5 @@
 [flake8]  # autopep8 would honor this setting, but not the other way around
-exclude = .eggs,*.egg,build,docs,.git,*cupyx/array_api,*tests/cupyx_tests/array_api_tests
+exclude = .eggs,*.egg,build,docs,.git,*cupy/array_api,*tests/cupy_tests/array_api_tests
 
 [tool:pytest]
 filterwarnings =
