+++ tests/cupy_tests/array_api_tests/test_elementwise_functions.py
@@ -2,9 +2,9 @@ from inspect import getfullargspec
 
 from numpy.testing import assert_raises
 
-from cupyx.array_api import asarray, _elementwise_functions
-from cupyx.array_api._elementwise_functions import bitwise_left_shift, bitwise_right_shift
-from cupyx.array_api._dtypes import (
+from cupy.array_api import asarray, _elementwise_functions
+from cupy.array_api._elementwise_functions import bitwise_left_shift, bitwise_right_shift
+from cupy.array_api._dtypes import (
     _dtype_categories,
     _boolean_dtypes,
     _floating_dtypes,
